modified semantic of plugin manifest, renamed file names according to PEP 8, refactored relative imports according to PEP 328
