+++ rqt/rqt_py_common/src/rqt_py_common/extended_combo_box.py
@@ -30,7 +30,7 @@
 # ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 
-import qt_gui.QtBindingHelper #@UnusedImport
+import qt_gui.qt_binding_helper #@UnusedImport
 from QtCore import Qt
 from QtGui import QComboBox, QCompleter, QSortFilterProxyModel
